+++ R/missSBM-package.R
@@ -4,7 +4,6 @@
 #' \itemize{
 #'   \item{\code{\link{observeNetwork}}}{a function to draw a partially observe network from an existing, fully observed network according to a variety of sampling designs}
 #'   \item{\code{\link{estimateMissSBM}}}{a function to perform inference of SBM from a partially observed under various sampling designs.}
-#'   \item{\code{\link{smooth}}}{a function to smooth an existing collection of missSBM estimation, to avoid being trapped in local maxima.}
 #' }
 #'
 #' These function leads to the manipulation of a variety of R objects instantiated from some R6 classes, with their respective fields and methods.