+++ inst/test_dev.R/simusNMAR.R
@@ -7,8 +7,6 @@ library(viridis)
 ## network parameters
 n <- 200
 pi <-  c(1/3,1/3,1/3)
-epsilon <- .25
-delta   <- .4
 lambda  <- matrix(.25,3,3) + diag(3) * (.15)
 theta <- list(mean = lambda)
 
@@ -23,7 +21,7 @@ control <- list(threshold = 1e-2, maxIter = 50, fixPointIter = 3, trace = 0)
 ## Simulation
 nbrSimu <- 100
 
-future::plan("multisession", workers=10)
+future::plan("multisession", workers = 10)
 
 RES <- lapply(1:nbrSimu, function(i) {
   res <- future.apply::future_lapply(list_psi, function(psi) {
@@ -43,10 +41,10 @@ RES <- lapply(1:nbrSimu, function(i) {
   out <- MNAR$doVEM(control)
 
   res <- data.frame(
-    ARI = c(aricode::ARI(MNAR$fittedSBM$memberships,A$memberships), aricode::ARI(MAR$fittedSBM$memberships,A$memberships)),
-    MSE = c(sqrt(sum((MNAR$fittedSBM$connectParam$mean-theta$mean)^2)), sqrt(sum((MAR$fittedSBM$connectParam$mean-theta$mean)^2))),
-    Psi = c(sqrt(sum((MNAR$fittedSampling$parameters-psi)^2)), NA),
-    variant = c("MNAR", "MAR")
+    ARI = c(aricode::ARI(cl_init,A$memberships), aricode::ARI(MNAR$fittedSBM$memberships,A$memberships), aricode::ARI(MAR$fittedSBM$memberships,A$memberships)),
+    MSE = c(NA, sqrt(sum((MNAR$fittedSBM$connectParam$mean-theta$mean)^2)), sqrt(sum((MAR$fittedSBM$connectParam$mean-theta$mean)^2))),
+    Psi = c(NA, sqrt(sum((MNAR$fittedSampling$parameters-psi)^2)), NA),
+    variant = c("init", "MNAR", "MAR")
   )
   res$simu  <- i
   res$delta <- psi[2,1] - psi[1,1]
@@ -60,6 +58,6 @@ p <- RES %>% dplyr::select(-Psi) %>% pivot_longer(c(ARI, MSE), names_to = "measu
   group_by(variant) %>%
   ggplot() + theme_bw() + viridis::scale_colour_viridis(discrete = TRUE) + viridis::scale_fill_viridis(discrete = TRUE) +
   aes(x = factor(delta), y = value, fill = variant) +
-  geom_violin() + geom_point(alpha = 0.25, size = .25) + facet_wrap(~ measure, scales = "free")
+  geom_boxplot() + geom_point() + facet_wrap(~ measure, scales = "free")
 
 print(p)